# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "KNU 80th Festival API",
        "description": "2026 \uacbd\ubd81\ub300\ud559\uad50 80\uc8fc\ub144 \ub300\ub3d9\uc81c \uc6f9\uc571 \ubc31\uc5d4\ub4dc API \u2014 \uc138\uc158 \uae30\ubc18 \uc778\uc99d (JSESSIONID \ucfe0\ud0a4)",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/matchings": {
            "post": {
                "tags": [
                    "\ub9e4\uce6d Command"
                ],
                "summary": "\ub9e4\uce6d \uc2e0\uccad",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MatchingCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "\uc2e0\uccad \uc131\uacf5",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseMatchingRegisterResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "\uc911\ubcf5 \uc2e0\uccad \ub610\ub294 \ub9e4\uce6d \uc77c\uc2dc\uc911\ub2e8",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseMatchingRegisterResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "\ub9e4\uce6d Command"
                ],
                "summary": "\ub9e4\uce6d \uc2e0\uccad \ucde8\uc18c",
                "operationId": "cancel",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MatchingAuthRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/matchings/result": {
            "post": {
                "tags": [
                    "\ub9e4\uce6d Query"
                ],
                "summary": "\ub9e4\uce6d \uacb0\uacfc \uc870\ud68c",
                "operationId": "getResult",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MatchingAuthRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseMatchingResultResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/matchings/status": {
            "patch": {
                "tags": [
                    "\ub9e4\uce6d Command"
                ],
                "summary": "\ub9e4\uce6d \uc11c\ube44\uc2a4 \uc0c1\ud0dc \ubcc0\uacbd",
                "operationId": "updateStatus",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MatchingStatusUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseMatchingStatusResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "MatchingCreateRequest": {
                "type": "object",
                "properties": {
                    "instagramId": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0,
                        "pattern": "^[A-Za-z0-9._]{1,100}$"
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "MALE",
                            "FEMALE"
                        ]
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 4
                    },
                    "nationality": {
                        "type": "string",
                        "maxLength": 10,
                        "minLength": 0
                    }
                },
                "required": [
                    "gender"
                ]
            },
            "ApiResponseMatchingRegisterResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/MatchingRegisterResponse"
                    },
                    "error": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                }
            },
            "MatchingRegisterResponse": {
                "type": "object",
                "properties": {
                    "instagramId": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "MATCHED",
                            "UNMATCHED",
                            "CANCELLED"
                        ]
                    },
                    "registrationDeadline": {
                        "type": "string"
                    },
                    "resultOpenAt": {
                        "type": "string"
                    }
                }
            },
            "MatchingAuthRequest": {
                "type": "object",
                "properties": {
                    "instagramId": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0,
                        "pattern": "^[A-Za-z0-9._]{1,100}$"
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 4
                    }
                }
            },
            "ApiResponseMatchingResultResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/MatchingResultResponse"
                    },
                    "error": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                }
            },
            "MatchingResultResponse": {
                "type": "object",
                "properties": {
                    "instagramId": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "MATCHED",
                            "UNMATCHED",
                            "CANCELLED"
                        ]
                    },
                    "resultOpen": {
                        "type": "boolean"
                    },
                    "matchedInstagramId": {
                        "type": "string"
                    },
                    "instagramProfileUrl": {
                        "type": "string"
                    },
                    "messageKo": {
                        "type": "string"
                    },
                    "messageEn": {
                        "type": "string"
                    }
                }
            },
...
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "type": "object"
                    },
                    "error": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                }
            },
...
            "MatchingStatusUpdateRequest": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "OPEN",
                            "PAUSED"
                        ]
                    },
                    "messageKo": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    },
                    "messageEn": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    }
                },
                "required": [
                    "status"
                ]
            },
            "ApiResponseMatchingStatusResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/MatchingStatusResponse"
                    },
                    "error": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                }
            },
            "MatchingStatusResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "OPEN",
                            "PAUSED"
                        ]
                    },
                    "messageKo": {
                        "type": "string"
                    },
                    "messageEn": {
                        "type": "string"
                    },
                    "registrationOpen": {
                        "type": "boolean"
                    },
                    "resultOpen": {
                        "type": "boolean"
                    },
                    "registrationDeadline": {
                        "type": "string"
                    },
                    "resultOpenAt": {
                        "type": "string"
                    },
                    "pendingCount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "matchedCount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "unmatchedCount": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
        }
    }
}